I#27649: "New ticket in" button with fixed width
When using a different language than English, the translated string might
be longer or smaller than the length('New ticket in'), so it needs to stretch
dynamically instead of having a fixed size.

The source for this issue is on the fact that rudder skin - differently from
the other ones - set the width with a selector including submit buttons (that
is, "input[type=submit]") which causes it to take precedence over the existing
rule in base/nav.css.