GraphViz.pm does not depend on the libgraph.so, but rather the "dot" program
This allows `configure` to better detect when GraphViz will be able to
be installed smoothly.  It may cause additional dependencies on an
upgrade -- however, failing to install these dependencies will not
change RT's behaviour in any way, as the --enable-graphviz option merely
controls default dependencies.  Enabling and disabling of the feature is
done at startup time, based on the presence of the GraphViz module or
the $DisableGraphViz configuration option.

This also makes RT's `configure` no longer require a working gcc.