Derived values need to be escaped before being used in LIKE
Consider the case of an Organization that was changed from "foo corp" to
"foo_corp".  The validator looks for things that are LIKE
"fsck.com-rt://%" and are NOT LIKE the current LocalURIPrefix, which
contains the Origanization.  As _ is a special character in LIKE
(matching any one character) the validator will not detect the old
invalid "fsck.com-rt://foo corp/..." entries, as they are indeed LIKE
"fsck.com-rt://foo_corp/%"

Explicitly escape all ocurrences of schema and prefix, in case they
contain % or _ characters.