Properly merge global defaults in to empty-hash homepages
The "Reset to default" button on the "RT at a Glance" RT user admin page
sets the preference to {}, not undef.  This worked prior to 9582e25a
because ->Preferences( 'name', { thing => "Default" }) merges in the
defaults into the hash, even if the hash exists.  This is different from
the behavior that 9582e25a assumed, where the hash is a straight-up
fallback which only kicks in if the value is undef.  This caused blank
hompages for users whose preferences had been reset.

Switch back to passing the global defaults to Preferences, which allows
for per-key merging.

Fixes I#30106.

(cherry picked from commit 0d0a348048a2e2654df71c883fd7ee967ccbe7df)