Handle enabled scrips in 4.1.5 upgrade, and disabled scrips in 4.2.2
90ba5667 rightly noticed that the 4.1.5 upgrade step had not adjusted
the Template column of disabled scrips.  In addition to adding a 4.2.2
upgrade step to handle them, it additionally changed the 4.1.5 upgrade
step to process them as well.

For installs which had already upgraded to 4.2.0 or 4.2.1, the change to
the 4.1.5 upgrade step was immaterial, and the missed scrips were fixed
by the 4.2.2 upgrade step.  However, it introduced a bug for installs
which were still on 4.0.x: namely, the 4.1.5 upgrade step would now
handle all scrips (including disabled scrips), followed the 4.2.2
upgrade step which would examine the disabled scrips -- and produce
warnings because they already contained characters.

One possibility is to remove the warning in the 4.2.2 upgrade step for
characters being found in the Template column; however, this might mask
problems in users on 4.2.0 or 4.2.1 for whom this is the first
examination of disabled scrips, and such a case is indeed a potential
problem.

Instead, opt to revert the 4.1.5 upgrade to only handling non-disabled
scrips.  The disabled scrips are thus handled identically by the 4.2.2
upgrade step for upgrades from both 4.0.x and 4.2.0/4.2.1.

This partially reverts commit 90ba56675bc5f288d4483f079f89a3b4c0d4cdc7.