allow default values for txn cfs
I got an interesting cache issue when implementing this:

considering that a txn's ObjectType is RT::Ticket, as we load the ticket object
in RT::Transaction::UpdateCustomFields(by calling $self->Object), it also
creates a ticket object cache(e.g. in ticket creation stage), which breaks
t/web/ticket_owner.t. in that test, the code is like:

    my $ticket = RT::Ticket->new( $user_a );
    my ($id, $txn, $msg) = $ticket->Create(
        Queue => $queue->id,
        Subject => 'test',
    );
    ok $id, 'created a ticket #'. $id or diag "error: $msg";
    is $ticket->Owner, RT->Nobody->id, 'correct owner';

    $agent_a->goto_ticket( $id );
    $agent_a->follow_link_ok({text => 'Take'}, 'Ticket -> Take');

    $ticket = RT::Ticket->new( RT->SystemUser );
    $ticket->Load( $id );
    ok $ticket->id, 'loaded the ticket';
    is $ticket->Owner, $user_a->id, 'correct owner';

so we first create a ticket in the test script process(which also creates a
cache), then take the ticket via web ui(which belongs to another process).
after that, in the test script process again, since there is a cache, the last
owner test fails(it loads the cached one).

I got around this by explicitly clearing the cache in
RT::Transaction::UpdateCustomFields to get back compatibility.