Fix unsetting of core date fields
b4c54faf switched to using ->IsSet instead of checking ->Unix to
determine if a date was set.  However, two of the changed tests (in
RT::Interface::Web) actually check "defined $date->Unix" -- which is
always true.  This led to it being impossible to unset core date fields.

Remove ->IsSet checks entirely, as the appropriate replacement for
"defined $date->Unix" is nothing, as it is always true.

Fixes: I#30180