DryRun now wraps in a rollback, and records all transactions
DryRun was previously limited to only running correspond or comment
transactions.  Change it to take a block of actions to run, which it
wraps in a transaction and rolls back at the end.  By additionally
tracking all transaction objects are created on the ticket, more changes
can be simulated, and TransactionBatch changes can be found.

Note that only transactionbatch scrips directly triggered by user
changes are visible -- as scrips are not committed, TransactionBatch
changes which would be recorded because of those indirect changes will
not be visible.