Don't ask for every field if you only want seven
When 'rt ls' is run in prettyprint mode (the default with no additional
arguments, or no specified fields), it flips on the long format, which
goes and gets a ton of fields, including, all the custom fields on each
ticket. If you have a lot of open tickets, this adds up to extra data
fetches and extra data over the wire which is then thrown away by
prettyprint().

Additionally, this is terribler on RTIR, which has some slow ACL checks
around custom fields and Queues, and the long format gets all the CFs
for each ticket. There is caching, but clearly it's not enough if you
have 100 open tickets. Relevant commits from rtir.git:1f5fb2a9 but that
cache still isn't enough. We'll do more digging in RTIR, but this takes
a list of a 100 RTIR tickets from minutes to 2 seconds.