Add squelched recipients as preference
User preference to start with squelch checkboxes unchecked/checked on
ticket reply page. This should respect the semi-permanetly squelched
address when applying its logic. If set to true, then the page shows
the checkboxes unchecked - which means the email addresses are *not*
getting any kind of mail.
Additionally, it adds a new config option called $SquelchedRecipients
which allows to set it broadly for all users.