Expose DisplayPath in ShowUpdateStatus
RTIR uses this to show the Unread Status on Incidents/IRs/etc
but RTIR tickets live in /RTIR/Display.html.  This linked to
/Ticket/Display.html?id=7#txn-id but we couldn't get and retain the
txn-id easily during the redirect to /RTIR/Display.html.  Instead, let
RTIR or another extension change where the link points on certain
tickets.

Part of the solution for issues #17450