Skipping ACL checks yields a sizable performance increase
The indexer spends quite a bit of time checking that the attachment
content is visible to the current user before indexing it -- all of
which is unnecessary, as it is run as the system user.  Explicitly
disable the ACL checking for the indexer, which removes a large number
of queries and yields a correspondingly sizable increase in indexing
speed.