Improve MySQL insert speed by batching inserts into one statement
MySQL must flush buffers after every insert statement; as such,
providing large numbers of INSERT statements is quite inefficient, as
most time is spent in disk I/O.  Instead, store rows to be inserted, and
insert them batch-by-batch.

This technique is not applicable to PostgreSQL because failure of the
to_tsvector call to convert a string would abort the entire insert.
Additionally, most installs use an additional column on the existing
table, which requires an UPDATE, and not an INSERT, which is not easily
batched.