Inversion of control of main indexing loops
Rather than having database-dependent if-statements, followed by a
standard loop which contains further database-dependent if-statements,
instead turn the loop into a function which cane be called from one of
two database-specific functions.  This is important because MySQL's
iteration will further diverge from PostgreSQL's in the following
commits.

This commit is best viewed with:
  git diff --patience --ignore-all-space