Indexing of "text" may fail for content with invalid UTF-8 byte sequences
Prior to f04f561f, content that claimed to be UTF-8 was inserted into
the database verbatim.  This errored immediately upon insertion for
PostgreSQL, as the Content column is of type "TEXT"; MySQL, however,
did not attempt to validate the bytes, as its Content column is of type
"LONGBLOB".

This thus only causes errors when attempting to insert the characters
into a "TEXT" column.  As multiple rows are inserted in batches, use the
row number from the error message to determine which row to blank, and
re-attempt insertion.