Still delay loading when showing headers
In 75a405b5 we added the ForceShowHistory flag to Click to load history
so that even if your JS was off, clicking on the link would load the
page.

This flag is also found in the Show Headers links added in 1a3327ce
and I think added so that when you clicked on ShowHeaders with Click to
load enable, you didn't *then* have to click on "Load history" just to
see the headers.

However, with delay, we already passed ShowHeaders to the helper, but
clicking on Show Headers forced us to do the page load inline instead.
This negates all of the perceived speed boost from delayed loading.
Removing it from the delay case hasn't immediately broken things, and
the delayed loading really doesn't play well with having JS off anyway.