Allow extra arguments to be passed to ShowHistory
RTIR wants to customize a number of paths during history display (see
RTIR/Display.html and RTIR/Incident/Display.html).  In order to use
delayed history loading, RTIR either needs to pass those as GET
arguments (allowing user injection of paths that we use to build links,
which is dangerous without sanitation) or otherwise get them to
ShowHistory.

With this, RTIR can just callback in the custom arguments (you can also
clobber any of the other arguments to ShowHistory with it).