Make TSVExport work on non-Tickets collections
a8232c75 attempted to refactor the existing /Search/Results.tsv into a
generic /Elements/TSVExport which took arbitrary collection objects.
However, it neglected to pass the class of the collection down through
to the ColumnMap code, causing it to only work on Tickets, or things
like CustomFields which are generally applicable across object classes.

Intuit the class as necessary, and pass it down through to the ColumnMap
code.