Fix dependency analysis of Scrips, ObjectScrips, and Queues
The changes to apply Scrips to multiple Queues were never brought into
the serializer.  Attempts to serialize a database with --scrips (and
without --clone) led attempting to call ->QueueObj on a RT::Scrip
object, which no longer has such a method; this led to the error:

    Can't locate object method "Id" via package "No object mapping for
    field"

Rework the dependency analysis of Scrips to mirror that of CustomFields,
in being many-to-many.