Allow modification of which types of links are shown
There's a callback to add data at the end of the Links portlet, but
nothing to allow you to change what's display.  This provides a callback
that would let you remove (for example) the Parents from the listing, so
that you can list them in a separate portlet later, with a custom
display.