Don't create empty LinkValueTo/IncludeContentForValue attributes
dec257a fixed RT::CustomField->Create() to accept LinkValueTo and
IncludeContentForValue, which are stored in attributes.  However, it
also ensured that those attributes are created even if they're empty.
This means every CF in the system, regardless of if it uses them, adds
two attributes.

Adjust _URLTemplate to remove attributes that are now empty, and more
explicitly return undef when the attribute does not exist.

Fixes: #22191