Allow the rt client to modify the status of a ticket while commenting
A reasonable use case for rt is to comment on a ticket (or
"correspond", which is just a comment that is also sent to the ticket
requestor).  In the web form for commenting, there is the ability to
update the status of the ticket at the same time.

This feature should be exposed in the command line client as well.

Bug-Debian: http://bugs.debian.org/760292