Ensure that all menus have href=""s, for screen readers
Screenreaders navigate menus by tabbing between them -- however, the
lack of an href="" on the "Actions" and "Feeds" means that they are not
valid tab targets.  As such, their sub-menus are not accessible.

Ensure that all menu items have an href, even if it is logically empty.
This ensures that screenreaders can reach it, and its sub-menus.