Improve paging for large resultsets
Our current solution works fine, except for the use case of having a large
resultset and so having a lot of page numbers to list. For instance, it
makes difficult to the user being on page 5 of 100 and wants to quickly
jump to page 55.

We pondered on having a box for one typing "55" to jump there, but although
that is the most functional solution, we think it's a awful UX.

Fixes #30374.