Add the "role" option to the preferences list
b12af0df added a new "role" user formatter; however, it neglected to add
it to the list of user formatters in RT::Config.  This means that if the
system default was changed to not be "role", users cannot switch back to
it.

Add the missing "role" as a possible preference for UsernameFormat.