Remove the unnecessary "on merge" lifecycle property
Lifecycles inexplicably have an on_merge property.  Since a merged
ticket cannot be found via searches or charts, its status is irrelevant.
The original change to set the status of a merged ticket was 30bbc418,
which provides no rationale for doing so.

Remove the extra setting, which merely serves to add unnecessary
complexity to lifecycles.