Set sessions to on-disk when MaxAttachmentSize is lowered, for Oracle
Oracle uses the MaxAttachmentSize to also control the maximum piece of
any BLOB column which is returned.  This hence also affects sessions,
and causes the attachment truncation tests to fail because login does
not succeed because the returned session data is truncated.

Set the test to use on-disk sessions to work around the spurious logouts
caused by a too-low MaxAttachmentSize on Oracle.