Fix the 4.2.7 upgrade to run on Oracle (and copy to 4.2.8)
The "Content" column of the Attributes table is a CLOB column on Oracle;
this makes it impossible to use the '=' operator on it.  Instead, check
the length of the contents, which works across databases and column
types.

The upgrade step is copied to 4.2.8, though only enabled on Oracle, to
catch those users who upgraded to 4.2.6 and for whom the upgrade step
errored out.  For Oracle users who upgrade from before 4.2.6 up to
4.2.8, the 4.2.8 step will be a no-op.

Fixes I#30391.