Allow ModifyTicket to change nobody -> someone else, without OwnTicket
Previously, CurrentUserCanSetOwner required OwnTicket if the ticket was
currently unowned, no matter if you were taking it yourself or assigning
it to someone else.

Allow nobody -> self with OwnTicket and (ModifyTicket, ReassignTicket,
or TakeTicket), and nobody -> someone else with ModifyTicket or
ReassignTicket.  This also closes a bug where TakeTicket and OwnTicket
was enough to change the owner from Nobody to someone else.

Fixes I#30406.