Apply squelching information to updates with no messages
Updates from the ticket Reply/Comment page which do not record content,
merely change the status / owner / etc., do not currently respect the
squelching checkboxes, despite showing as if they do.  This is because
the squelching information is stored on the Correspond/Comment
transaction, and thus only propagated to other transactions if such was
created.

Even on updates that do not have content, store and respect the
squelching information.  This builds upon 4a33c662, which made the
transaction squelching settings apply for the lifetime of the ticket
object.