Simplify WhoHaveRoleRight and WhoHaveGroupRight queries before unioning
Perform a true union in SQL, via a separate 'in' query.  This
effectively prevents duplicates in the resultset (which
DBIx::SearchBuilder::Union makes no attempt to do), orders the set as a
whole, and allows the individual queries to be simpler.  In particular,
the lack of "DISTINCT" and "ORDER BY" on the WhoHaveGroupRight query on
MySQL causes it to generate a query plan which no longer includes "with
temporary table" (due to the DISTINCT) or "with filesort" (due to the
ORDER BY), and makes better use of indexes (due to the limited number of
columns returned).

This does come at the assumption that duplicates within each of the two
queries are not common -- that is, that users will not have the relevant
right by way of a large number of membership paths.