Forwarded messages break attachments with non-ASCII filenames
On RT 4.2.3, forwarded messages lost header info cause RT takes every
header from DB and apply MIME transfer encoding to each as a full line,
thus losing information about the internal structure of the header, if
it has any.

This aim to fix by applying the correct encoding according to rfc2231
before assembling the MIME object. It worths mentioning that we are not
using MIME::Field::ParamVal to assemble the parts, because at the moment
it doesn't know how to do so correctly.

Fixes: #29753.