Our rich text editor produces entities for indexed characters
Postgres is completely capable of searching for cherché but the rich
text editor produced cherch&eacute; and so we handed that to Pg and it
stored 'cherch' which won't match later.

This causes us to pass the actual word to Pg to be tokenized and stored
improving the odds that a later search will find it.

Since RT generates text/plain outgoing mail parts and those parts
generally contained the decoded word, this may have been hiding this
failure (the search would find the outgoing mail message rather than
the reply itself, but would still match the ticket).