Switch interactive tools to use Init()
Scripts which may be run from cron jobs and the like are not altered; as
3ce1ac4e notes, the use of Init() may cause additional warnings to be
logged.  While more correct, such a change is incorrect for a stable
series; thus, only manually-run upgrade steps and command-line tools are
altered.

This also fixes a bug, introduced in 0d1cd53e, which caused the upgrade
steps to no longer always log to STDERR.  In switching to:

    use RT -init;
    RT->Config->Set(LogToSTDERR => '...');

The latter command did not alter the effective logging settings, as
RT::Init had already been called.

This commit changes the upgrade steps use Init(), and thus to default to
displaying only "warning" levels to the user; the --verbose option is
now available if debug output is still desired.