Forward-port run_validator (and timeout functionality) from RT 4.0
547af3b4 added a test to check that rt-validator didn't infinite loop;
as such, it added timeout functionality to run_validator in RT::Test.
run_validator had been extracted in 21c7ddf7; however, the _same_
extraction was also done on 4.2-trunk in 80ae47db, but placed the
function elsewhere in the file.  When 4.0-trunk was merged to 4.2-trunk
in 1f4a0bc7, it (without a merge conflict!) placed a duplicate into
RT::Test -- which was removed by hand before being committed.

Unfortunately, the version that was removed was the 4.0-trunk version,
which had the timeout feature added in 547af3b4.  As such, 4.2-trunk's
test has since been passing --timeout=30 to rt-validator, instead of
setting a 30-second alarm.  Because rt-validator did not validate its
command-line arguments, and the code did _not_ infinite-loop, this went
unnoticed.

9d354ba switched rt-validator to use RT::Interface::CLI::Init(), which
does error if command-line argument parsing failed, thus exiting with a
non-zero exit status, failing the test and making this situation clear.

Forward-port the run_validator from 4.0-trunk, giving it the
functionality intended to be added in 547af3b4 by way of 1f4a0bc7.