Pass down the CustomField object from the collection.
Given a RT::CustomFieldValue object you can now do the same
introspection that External Custom Fields can do on
RT::CustomFieldValues by looking at the actual CustomField object that
was used to load the collection.  This is transparently returned by
->CustomFieldObj as long as SetCustomFieldObj was called previously and
passed a CustomField which matches the one the database says this
CustomFieldValue belongs to.