Allow Groups be added as Requestor
The "Scope" argument is used to determine if we're trying to list the
watcher types we can change on queues or anything else -- which for the
later case includes "Requestor" besides the common "AdminCc" and "Cc"
types.

This came about in 6020dac5, which is a famously unhelpful giant lump
commit from the 2.1 days, so it doesn't have any justification provided
in the commit message for why the groups are treated differently than
the queues.

Thus, it removes the "Scope" param in order to fall into default case
as for the user's call.