Always allow role and system groups to be enumerated
Previously, the SeeGroup right controlled RT::Groups results even for
internal groups; this caused the queue rights page for a user with
AdminQueue but not SeeGroup to not list Everyone / Privileged /
Unprivileged, nor the queue role groups.

Allow system groups to always be seen, and role groups to be seen if the
user can see the object the role group is on.  This is a broadening of
the privileges that previously existed.

Note that as ->CurrentUserCanSee is not used to ACL ->_Value, this does
not quite enforce a group ACL globally; any explicitly loaded group can
still be examined.

Fixes: I#30416