Always allow role and system groups to be enumerated
Previously, the SeeGroup right controlled RT::Groups results even for
internal groups; this caused the queue rights page for a user with
AdminQueue but not SeeGroup to not list Everyone / Privileged /
Unprivileged, nor the queue role groups.

Allow system groups to always be seen, and role groups to be seen if the
user can see the object the role group is on.  This is a broadening of
the privileges that previously existed.

Note that as this limit applies to ->Next, and not ->AddRecord, and as
->CurrentUserCanSee is not used to ACL ->_Value, this does not enforce a
group ACL globally.  ->ItemsArrayRef will still return all matching
groups, regardless of rights, and any explicitly loaded group can be
examined.

Moving the ACL to ->AddRecord is complicated at this time, as users need
the ability to see groups which are watchers on tickets they can see,
which is difficult to impose in a performant manner.

Fixes: I#30416