Add X-RT-Interface: Email to new comments and correspondence
9189e698 introduced the X-RT-Interface: header, to allow for automated
detection of the source of a comment or correspondence.  However, the
header was only explicitly set for incoing email during ticket creation,
not during updates; as such, updates were marked as X-RT-Interface: API.

Ensure that the header is added to all incoming email, not just those
that take the Create path.

Fixes: I#30388