Switch from the DBA-only tables, to tables the user can inspect
Despite the upgrade process being run as RT's "dba" user, in many Oracle
installs the dba user is set to the RT user, which is granted full
rights on its tablespace.  However, the dba_* tables are not accessible
to this user.

Switch to using the all_indexes and associated tables, which only
include the data which is accessible to the user, unlike the dba_indexes
table, which is entirely unreadable by users which are not truely the
DBA.

Fixes: I#30393