only bars need to be lifted up a bit
bar images have a padding-top like stuff themselves, I believe that's why we
lifted the whole bar image up so they can be top aligned horizontally with the
data table.

pie images don't have this issue and if we lift them up too, they could
overlap with the div.titlebox.title in a dashboard when data table is not
shown.