Properly walk the Queue parts of the TicketSQL in GetReferencedQueues
Previously, GetReferencedQueues simply OR'd the positive Queue
references of the query together.  This did not take into account ACLs,
negative restrictions, or AND'd limits.

Walk the parsed query tree, removing nodes that do not reference Queue
limits, and removing empty AND/OR nodes; this produces a tree which
details which queues can possibly match.  Applying this tree to a
RT::Queues object produces ground-truth on what queues the matching
tickets can possibly be from.

This information can be used to trim the Owners and Status drop-downs
even when no queue have been selected.  For backwards compatibility, the
set of possibly-matching CFs is only shown if a Queue limit is found.