Expose Article's Topics.html to SelfService
Topics.html is pretty simple, so as long as permissions are set (needs
SeeClass and ShowArticle) it will render just fine for an
Unprivileged user.  As such, shim it into /SelfService/.

Unfortunately, only one of the links in Topics.html was relative, but if
you make the rest relative, it'll work great in the Privileged view
*and* SelfService without copying anything.  It does mean Topics.html
lives in /SelfService/ rather than /SelfService/Article/ but I think
that's ok, especially given the /Articles/Article/ hierarchy found
through the Privileged view.