Load a Custom Field to improve searches
Passing the name causes RT to generate a query that joins through
CustomFields, ObjectCustomFields, ObjectCustomFieldValues and performs
really poorly on anyone with real data in OCFVs.

If you pass in an id, LimitCustomField will load as SystemUser and
create better SQL (see a4c8bfa4 for more).  If you pass in a real
object, there's even more optimizations the code can make.

Loading as SystemUser is concerning, however the Fields to search by are
pulled from the system configuration and are not tweakable by any of the
User endpoints.  Someone could write code that calls
RT::Users->SimpleSearch, but that's outside core.

Code using this system CF object are just inspecting values, so should
be safe.

The code has always walked around ACLs on the User CF for searching
becaues it joined to CustomFields by name, without loading the object
and sanchecking if the currentuser can see it.  This maintains backwards
compatibility with the behavior, since an admin saying "Search this CF"
should continue to just work.

If configured with a User CF it cannot load (or if someone calls this
method on an RT::Users collection) we will not generate a poorly
performant search and instead will warn and skip the CF.